Adding new output type - Javascript Object
I love the fact that the MySQL output can be pasted into SQL editor. I needed the same thing for pasting into Javascript editor when dealing with really large spreadsheets. These changes will allow users to paste the data output straight into their Javascript editor; providing all of the benefits of working with a native Javascript object (and all of its properties).
The code changes in this commit are simply the addition of native instantiation of a Javascript object named mrDataConverterObj, the removal of the newline character (after each record) and enclosing the entire output with single quotes (as you would do if you wanted to parse the output as a JSON object in Javascript).